R3 AR-1: extract FFI entry points into a submodule
Move the C-ABI parser_override plumbing out of parse/mod.rs into
parse::ffi:

- OverrideContext (the empty FFI-shape marker) and its lifecycle exports
  sv_make_override_context / sv_drop_override_context.
- The parser-override hook sv_parser_override_rust and the parse-function
  validation hook sv_parse_function_rust, plus the run_validation_for_
  parse_function bridge (both the extension and the syntax-only test
  variant) and the publish_owned_sql / sv_free_buffer buffer helpers.

This isolates all the unsafe C-boundary code from the pure parse logic.
The #[no_mangle] symbols are unaffected by the module move (the linker
sees the same names). OverrideContext and sv_free_buffer are re-exported
from mod.rs under `#[cfg(feature = "extension")]` so crate::parse::*
paths (native_sql's OverrideContext use, and the ddl::alter_helpers_ffi
test's sv_free_buffer call) resolve unchanged; the sv_* hooks used by the
in-module tests are re-exported under matching test cfgs. write_error_to_
buffer's import returns to mod.rs under #[cfg(test)] for the two
truncation unit tests that remain there.

No behaviour change. Verified: just build (extension), cargo test (1045),
clippy -D warnings, and just test-sql (64/64) all pass.
